agregando funcion de pago y envio de codigo via email

+++ src/main/java/com/vent_entra_sf/proyect_vent_sofi/dto/compra/CompraCreateDTO.java
@@ -15,5 +15,5 @@ public class CompraCreateDTO {
     private String apellido;
     private String email;
     private Long eventoId;
-    private BigDecimal cantidadEntradas;
+    private Integer cantidadEntradas;
 }
